requirements.txt
datasets==2.9.0
Pillow==9.4.0
streamlit==1.17.0
torch==1.13.1
transformers==4.26.0
